requirements/project.txt
mercurial==2.0
Django==1.4.1
#PIL==1.1.7
South==0.7.6
django-appmedia==1.0.1
django-classy-tags==0.3.4.1
django-cms==2.3
django-debug-toolbar==0.8.5
django-indexer==0.3.0
django-mailchimp==0.1.27
django-mptt==0.5.2
django-paging==0.2.4
django-reversion==1.6.2
django-sekizai==0.6.1
django-sentry==1.8.6.2
django-snippetscream==0.0.5
django-templatetag-sugar==0.1
#-i http://pypi.python.org/simple/
hg+https://code.google.com/p/gdata-python-client/@e034a847a474459906cd945688826747a4c13575#egg=gdata-dev
#psycopg2==2.4.2
Jinja2==2.6
#eventtools
git+https://github.com/GaretJax/coffin.git@8e2bd87fa9fb4ad9d4b3d2c49777c51ebd0b14a6#egg=Coffin-dev
html5lib==0.95
python-dateutil==1.5
vobject==0.8.1c
